The major change is that the Sigma(E) calculation algorithm was fixed on a couple of accounts. The finalStateIndex variable is now properly tracking the right final states.
I also modified the bondAnalysis script to slightly increase the default
distance between components when the system is turned into an openscad model.
This is a hard coded value and it should probably be intelligently selected
in the future.

I also fixed a problem with the shifting of the kpoint mesh. It had been
failing to shift and now it should shift according to the space group. This
was in makeinput.

The other changes were all not errors and just improved readability.