So basically a ton of changes have been made. The key ones are:
(1) All of the modules in the program now make use of the "use, only:" method
to selectively pull in parts of a module. The point here is to make it easier
for readers of the program to find out where variables and subroutines come
from. (E.g. if you are reading some subroutine and a variable is used or
another subroutine is used and you want to know where it comes from, then you
can look at the list of modules at the beginning of your subroutine and it
will list everything that is pulled from everywhere else. If you don't see it
in your list, then either the variable/subroutine is inside your current
module with you, or it is associated with a module use command that does not
have the "only:" option. (The second case only happens about twice in the
whole codebase and it happens only because *everything* was used from that
module.)

(2) A number of new scripts have been added. (A) dump2skl is a good foundation
for this general process, but it doesn't work in the general case yet. It is
only usable for the specific a-BC:H model system. I believe that the extension
to a more general form will be relatively easy. (B) The massGather and massPrep
scripts are a bit crude tools to help prepare input for executing a large set
of olcao calculations on a parallel system. They need considerable refinement.
(C) The modStruct program has *finally* been ported to this version of OLCAO
after a long (multi-year) wait. This program is handy for creating lots of
modifications to models (making grain boundaries, surfaces, etc. rotating
molecules, making supercells, perturbing structures, etc.)

(3) The compile system was also "improved" to take advantage of the AUTOMAKE
subdir-objects option in the configure.ac file. (This was basically forced on
us because of a change in the way that the autotools behaves with respect to
a source code directory tree with Makefile.am files at each level.) Anyway,
hopefully this will be better in the end. However, I am currently not
satisfied with it because of an annoyance with the way that it seems to treat
the compilation of the general and gamma k-point versions of the program.
Essentially, if I change a file then in principle any file that depends on it
should be relinked. The problem is that it does not appear to be "finding" the
right file to attempt to link with. I.e. a file which must be preprocessed
because it has gamma k-point and general k-point components will need to
relink to form gamma and general k-point executables. Unfortunately, the
make operation seems to be grabbing the wrong other files when attempting to
relink. So, when trying to relink a gamma k-point version of setup, it grabs
the general k-point version of the code (which obviously will not contain the
right subroutines and variable etc.)

(4) In the .olcao file you will note that the old variables used to identify
the queue submission method have been consolidated into OLCAO_QUEUE which can
have different values. Note that one effect of this is that in the olcao script
the execution operation now has a prefix associated with the particular
submission method. (E.g. slurm systems require a "srun -n1 --exclusive" prefix
for olcao serial jobs.)

(5) Probably a bunch of other small things.