So this commit has one change to the serial OLCAO. The makeinput script was modified to better deal with making choices for the symmetric band structure path to follow when doing sybd calculations.
The main modification is the addition of the polcao directory and a bunch of
files in it that pulled from Jim Currie's branch. This polcao is not yet
operational, but it includes a number of important steps along the way.

Most of the global arrays stuff has been removed. I have mostly merged the
intg and band programs into one. The setup program is mostly parallelized.