In this commit we have:
(1) Changed the build system from GNU autotools to cmake because autotools
was not working well with Fortran 90 module files. The cmake program is
working *much* better.
(2) The momentum matrix element calculation in gaussIntegrals.f90 has been
reverted to the expressions from a few versions ago because it was failing
to compute correct values when compiled with gfortran. This will have to be
investigated more in the future.
(3) There were a few small changes regarding formatted output so eliminate
some warnings about the size of the field vs the number of digits shown after
the decimal point.