A bunch of changes were made.
The most important one is that a preliminary implementation of LDA+U has been
created. This required a lot of changes including to the makeinput script, the
format of the scfV.dat and its related files.

I have added Ben Walker's numerical integration routines for checking the
accuracy of the analytic integration formulas. I also added the scripts that
he wrote for generating the analytic formulas.