A few significant changes:
I changes the CMakeLists.txt file to make OLCAO compile correctly with the
ifort compiler. Optimizations are turned on except that arrays are all going to
be heap allocated arrays instead of the default stack allocated. If this is not
done, then the program will die when a "too large" array is allocated.

I changed the potentialUpdate.f90 file to extend the potential mixing code to
more than a depth of two. Note that as mentioned in the D.G. Anderson paper,
it is probably not beneficial to go beyond a depth of about 5.

I also put the memory estimation code back into makeinput. Just as before, this
code produces a crude estimation. (In my mind a bit too crude.) It needs to be
studied and compared to actual data and then fixed. However, it is a start and
it follows the perl strict pragma now. (It did not follow the strict pragma in
the past and that is why it was never carried over until now.)