A number of minor changes were made and a couple of more significant changes were also made. The minor changes include changes to array indices to try to prevent array (or matrix) copies from being generated when they are passed between subroutines. A bunch of comments were added or improved.
More significantly:
(1) The main hdf5 file was extended so that the potential data from every
iteration is recorded. (No additional scripts have yet been written to aid in
extracting that data from the hdf5 file conveniently, but at least the data is
there now.)
(2) The potential update feedback level can be extended to arbitrary numbers
now. Previously the options were 0, 1, and 2. It has been found that a 4 or 5
provides an improvement while higher than that appears to be of limited value.
(3) The possibility of using minimization of the total energy instead of the
root of the SCF potential variation function has been introduced. This still
needs to be tested more fully.