Fixed an error in bond.F90. It now (should) correctly incorporate variable charge populations in states when producing Q* and bond order data.
Introduced a check directory with some precomputed results. The idea is that
when someone compiles OLCAO on a new machine, they can check against these
results to determine if their new code is running correctly. (This will be
improved over time.)

Threw in two completely unrelated scripts that I use to help manage attendance
in my classes. I'll probably remove them and make a separate repository at a
future date, but for now this was the fastest way to put them into version
control.

Added the preliminary ensembleCondense (which does not directly use OLCAO) but
it does use some of the existing scripts.