Fix issues with the makeFittedRhoV script to adapt it to the new file format for the potential function.
Also, correct the Boron isolated atom potential file to be consistent with the
elements.dat specification for the minimum potential term. It should be 0.3
and not 0.2. In the OLCAO book it is specified as 0.2, but I believe that is
a typographical error and that the currect value should be 0.3. (That number
is what was used in much prior research.)